[chore] adds compile time checks for unsafe pointer casting (#4576)
Reviewed-on: https://codeberg.org/superseriousbusiness/gotosocial/pulls/4576
Co-committed-by: kim <grufwub@gmail.com>

+++ internal/router/attach.go
@@ -21,6 +21,7 @@ import (
 	"slices"
 	"unsafe"
 
+	"code.superseriousbusiness.org/gotosocial/internal/gtserror"
 	"github.com/gin-gonic/gin"
 )
 
@@ -34,6 +35,10 @@ func (r *Router) AttachNoRouteHandler(handler gin.HandlerFunc) {
 }
 
 func (r *Router) AttachGroup(relativePath string, handlers ...gin.HandlerFunc) *RouterGroup {
+	if unsafe.Sizeof(RouterGroup{}) != unsafe.Sizeof(gin.RouterGroup{}) ||
+		unsafe.Offsetof(RouterGroup{}.RouterGroup) != 0 {
+		panic(gtserror.New("compile time unsafe pointer assertion"))
+	}
 	handlers = slices.DeleteFunc(handlers, func(h gin.HandlerFunc) bool { return h == nil })
 	return (*RouterGroup)(unsafe.Pointer(r.engine.Group(relativePath, handlers...)))
 }
